fix: deployment bundle settings are not handled in server mode
