+++ src/AWS.Deploy.Common/Recommendation.cs
@@ -28,9 +28,11 @@ namespace AWS.Deploy.Common
 
         public DeploymentBundle DeploymentBundle { get; }
 
+        private readonly List<OptionSettingItem> DeploymentBundleSettings = new ();
+
         private readonly Dictionary<string, string> _replacementTokens = new();
 
-        public Recommendation(RecipeDefinition recipe, ProjectDefinition projectDefinition, int computedPriority, Dictionary<string, string> additionalReplacements)
+        public Recommendation(RecipeDefinition recipe, ProjectDefinition projectDefinition, List<OptionSettingItem> deploymentBundleSettings, int computedPriority, Dictionary<string, string> additionalReplacements)
         {
             Recipe = recipe;
 
@@ -38,6 +40,7 @@ namespace AWS.Deploy.Common
 
             ProjectDefinition = projectDefinition;
             DeploymentBundle = new DeploymentBundle();
+            DeploymentBundleSettings = deploymentBundleSettings;
 
             _replacementTokens[REPLACE_TOKEN_PROJECT_NAME] = Path.GetFileNameWithoutExtension(projectDefinition.ProjectPath);
 
@@ -78,6 +81,11 @@ namespace AWS.Deploy.Common
             }
         }
 
+        public IEnumerable<OptionSettingItem> GetConfigurableOptionSettingItems()
+        {
+            return Recipe.OptionSettings.Union(DeploymentBundleSettings);
+        }
+
         /// <summary>
         /// Interactively traverses given json path and returns target option setting.
         /// Throws exception if there is no <see cref="OptionSettingItem" /> that matches <paramref name="jsonPath"/> />
@@ -98,7 +106,7 @@ namespace AWS.Deploy.Common
 
             foreach (var id in ids)
             {
-                var optionSettings = optionSetting?.ChildOptionSettings ?? Recipe.OptionSettings;
+                var optionSettings = optionSetting?.ChildOptionSettings ?? GetConfigurableOptionSettingItems();
                 optionSetting = optionSettings.FirstOrDefault(os => os.Id.Equals(id));
                 if (optionSetting == null)
                 {
